2023-04-19 split project to components
