+++ src/components/App/App.scss
@@ -5,116 +5,13 @@
     background-color: #8d8d8d;
     margin: 2vh 0 1vh 0;
   }
-
-  &__loader {
-    margin-bottom: 1vh;
-  }
 }
 
 .opacity {
   opacity: 0.5;
 }
 
-// .app body html {
-//   position: relative;
-// }
-
 // ****************************************************************
-
-.header {
-  background-color: rgb(204, 204, 204);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 10vh;
-}
-
-.guests {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-
-  &__wrapper {
-    border-bottom: 5px solid rgb(117, 193, 255);
-    margin-bottom: 2vh;
-  }
-
-  &__list {
-    display: flex;
-    // justify-content: center;
-    // align-items: center;
-    flex-direction: column;
-    height: 50vh;
-    width: 70vw;
-    overflow-y: scroll;
-    background-color: rgb(245, 245, 245);
-    margin-top: 2vh;
-  }
-
-  .item {
-    &__button {
-      min-width: 8vw;
-      min-height: 3vh;
-      background-color: #8d8d8d;
-    }
-  }
-
-  .item--additional {
-    background-color: rgb(192, 192, 192);
-  }
-
-  &__item {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    min-width: 50vw;
-    min-height: 3vh;
-    background-color: rgb(231, 231, 231);
-
-  }
-
-  &__description {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    min-width: 50vw;
-    min-height: 3vh;
-    background-color: rgb(218, 218, 218);
-  }
-
-  &__form {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 1vh;
-
-    .form {
-      &__tariff {
-        text-align: center;
-        min-width: 10vw;
-        min-height: 3vh;
-        background-color: rgb(199, 199, 199);
-      }
-
-      &__input {
-        min-width: 15vw;
-        min-height: 3vh;
-        background-color: #dbdbdb;
-        padding-left: 10px;
-      }
-
-      &__button {
-        min-width: 8vw;
-        min-height: 3vh;
-        background-color: #8d8d8d;
-      }
-    }
-  }
-}
-
-
-
 .paymentModal {
   position: absolute;
   width: 100vw;
@@ -201,116 +98,5 @@
     margin: 0 3vw 0 3vw;
     font-weight: bold;
     font-size: 30px;
-  }
-}
-
-
-.lds-ellipsis {
-  display: inline-block;
-  position: relative;
-  width: 80px;
-  height: 80px;
-}
-.lds-ellipsis div {
-  position: absolute;
-  top: 33px;
-  width: 13px;
-  height: 13px;
-  border-radius: 50%;
-  background: rgb(97, 97, 97);
-  animation-timing-function: cubic-bezier(0, 1, 1, 0);
-}
-.lds-ellipsis div:nth-child(1) {
-  left: 8px;
-  animation: lds-ellipsis1 0.6s infinite;
-}
-.lds-ellipsis div:nth-child(2) {
-  left: 8px;
-  animation: lds-ellipsis2 0.6s infinite;
-}
-.lds-ellipsis div:nth-child(3) {
-  left: 32px;
-  animation: lds-ellipsis2 0.6s infinite;
-}
-.lds-ellipsis div:nth-child(4) {
-  left: 56px;
-  animation: lds-ellipsis3 0.6s infinite;
-}
-@keyframes lds-ellipsis1 {
-  0% {
-    transform: scale(0);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-@keyframes lds-ellipsis3 {
-  0% {
-    transform: scale(1);
-  }
-  100% {
-    transform: scale(0);
-  }
-}
-@keyframes lds-ellipsis2 {
-  0% {
-    transform: translate(0, 0);
-  }
-  100% {
-    transform: translate(24px, 0);
-  }
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-.loader {
-  transform: translateZ(1px);
-}
-.loader:after {
-  content: '$';
-  display: inline-block;
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  text-align: center;
-  line-height:40px;
-  font-size: 32px;
-  font-weight: bold;
-  background: #c9c9c9;
-  color: #e9e9e9;
-  border: 4px double ;
-  box-sizing: border-box;
-  box-shadow:  2px 2px 2px 1px rgba(0, 0, 0, .1);
-  animation: coin-flip 4s cubic-bezier(0, 0.2, 0.8, 1) infinite;
-}
-@keyframes coin-flip {
-  0%, 100% {
-    animation-timing-function: cubic-bezier(0.5, 0, 1, 0.5);
-  }
-  0% {
-    transform: rotateY(0deg);
-  }
-  50% {
-    transform: rotateY(1800deg);
-    animation-timing-function: cubic-bezier(0, 0.5, 0.5, 1);
-  }
-  100% {
-    transform: rotateY(3600deg);
   }
 }
