+++ src/components/Guests/Guests.scss
@@ -1,0 +1,54 @@
+.guests {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+
+  &__wrapper {
+    border-bottom: 5px solid rgb(117, 193, 255);
+    margin-bottom: 2vh;
+  }
+
+  &__list {
+    display: flex;
+    // justify-content: center;
+    // align-items: center;
+    flex-direction: column;
+    height: 50vh;
+    width: 70vw;
+    overflow-y: scroll;
+    background-color: rgb(245, 245, 245);
+    margin-top: 2vh;
+  }
+
+  .item {
+    &__button {
+      min-width: 8vw;
+      min-height: 3vh;
+      background-color: #8d8d8d;
+    }
+  }
+
+  .item--additional {
+    background-color: rgb(192, 192, 192);
+  }
+
+  &__item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    min-width: 50vw;
+    min-height: 3vh;
+    background-color: rgb(231, 231, 231);
+
+  }
+
+  &__description {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    min-width: 50vw;
+    min-height: 3vh;
+    background-color: rgb(218, 218, 218);
+  }
+}
