+++ src/components/Guests/index.js
@@ -1,0 +1,2 @@
+import Guests from './Guests';
+export default Guests;
