+++ src/components/GuestsAdd/index.js
@@ -1,0 +1,2 @@
+import GuestsAdd from './GuestsAdd';
+export default GuestsAdd;
