+++ src/components/GuestsLoader/GuestsLoader.jsx
@@ -1,0 +1,9 @@
+import './GuestsLoader.scss';
+
+function GuestsLoader() {
+  return (
+    <span class="guests__loader"></span>
+  )
+}
+
+export default GuestsLoader;
