+++ src/components/GuestsLoader/index.js
@@ -1,0 +1,2 @@
+import GuestsLoader from './GuestsLoader';
+export default GuestsLoader;
