+++ src/components/GuestsSearch/index.js
@@ -1,0 +1,2 @@
+import GuestsSearch from './GuestsSearch';
+export default GuestsSearch;
