+++ src/components/Header/Header.jsx
@@ -1,0 +1,11 @@
+import './Header.scss';
+
+function Header() {
+  return (
+      <header className="header">
+        <h1 className="header__title">Admin Moore</h1>
+      </header>
+  )
+}
+
+export default Header;
