+++ src/components/Header/Header.scss
@@ -1,0 +1,7 @@
+.header {
+  background-color: rgb(204, 204, 204);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 10vh;
+}
